fix: move base Lexer and fix javascript syntax
